requirements.txt
langchain
langgraph
fastapi
uvicorn
python-dotenv
beautifulsoup4
requests
pytest
faiss-cpu
bubble-client
langchain-community
langchain-google-genai
